+++ cmd/convert/convert.go
@@ -118,7 +118,7 @@ func CliConfig() *cli.Command {
 			var conversionStarted chan bool
 			var inputVideoCodecDetected chan mediaConvert.InputVideoCodecDetectedBatchMessage
 
-			var batchTask mediaConvert.BatchConverterTask
+			var batchTask mediaConvert.BatchTask
 
 			if c.String("config") != "" {
 				configFile := files.NewFile(c.String("config"))
@@ -144,18 +144,20 @@ func CliConfig() *cli.Command {
 				outFile := inFile.Clone()
 				outFile.SetDirPath(files.NewPath(outputPath))
 
-				batchTask = mediaConvert.BatchConverterTask{
+				batchTask = mediaConvert.BatchTask{
 					Parallelism: c.Int("parallelism"),
-					Tasks: []mediaConvert.ConverterTask{
-						mediaConvert.ConverterTask{
-							InFile:       inFile,
-							OutFile:      outFile,
-							HWAccel:      c.String("hwa"),
-							VideoCodec:   c.String("video-codec"),
-							Preset:       c.String("preset"),
-							VideoBitRate: c.String("video-bitrate"),
-							VideoQuality: c.Int("video-quality"),
-							Scale:        c.String("scale"),
+					Tasks: []mediaConvert.Task{
+						mediaConvert.Task{
+							InFile:  inFile.FullPath(),
+							OutFile: outFile.FullPath(),
+							Params: mediaConvert.Params{
+								HWAccel:      c.String("hwa"),
+								VideoCodec:   c.String("video-codec"),
+								Preset:       c.String("preset"),
+								VideoBitRate: c.String("video-bitrate"),
+								VideoQuality: c.Int("video-quality"),
+								Scale:        c.String("scale"),
+							},
 						},
 					},
 				}
@@ -163,17 +165,19 @@ func CliConfig() *cli.Command {
 				if c.Bool("recursively") {
 					outputPath := c.Args().Get(1)
 
-					batchTask, err = mediaConvert.BuildBatchTaskFromRecursive(mediaConvert.RecursiveConverterTask{
-						Parallelism:  c.Int("parallelism"),
-						InPath:       files.NewPath(inputPath),
-						OutPath:      files.NewPath(outputPath),
-						HWAccel:      c.String("hwa"),
-						VideoCodec:   c.String("video-codec"),
-						Preset:       c.String("preset"),
-						VideoBitRate: c.String("video-bitrate"),
-						VideoQuality: c.Int("video-quality"),
-						Scale:        c.String("scale"),
-					}, infoGetter)
+					batchTask, err = mediaConvert.BuildBatchTaskFromRecursive(mediaConvert.RecursiveTask{
+						Parallelism: c.Int("parallelism"),
+						InPath:      files.NewPath(inputPath),
+						OutPath:     files.NewPath(outputPath),
+						Params: mediaConvert.Params{
+							HWAccel:      c.String("hwa"),
+							VideoCodec:   c.String("video-codec"),
+							Preset:       c.String("preset"),
+							VideoBitRate: c.String("video-bitrate"),
+							VideoQuality: c.Int("video-quality"),
+							Scale:        c.String("scale"),
+						},
+					}, mediaInfoGetter)
 
 					if err != nil {
 						return errors.Wrap(err, "Building recursive task")