+++ pkg/media/convert/converter_models.go
@@ -26,27 +26,30 @@ type VideoFileFilteringMessage struct {
 	Err     error
 }
 
-// BatchConverterTask _
-type BatchConverterTask struct {
-	Parallelism           int             `yaml:"parallelism"`
-	StopConversionOnError bool            `yaml:"stop_conversion_on_error"`
-	Tasks                 []ConverterTask `yaml:"tasks"`
+// BatchTask _
+type BatchTask struct {
+	Parallelism           int    `yaml:"parallelism"`
+	StopConversionOnError bool   `yaml:"stop_conversion_on_error"`
+	Tasks                 []Task `yaml:"tasks"`
 }
 
-// ConverterTask _
-type ConverterTask struct {
-	ID               string      `yaml:"id"`
-	InFile           files.Filer `yaml:"in_file"`
-	OutFile          files.Filer `yaml:"out_file"`
-	VideoCodec       string      `yaml:"video_codec"`
-	CPUDecoding      bool        `yaml:"cpu_decoding"`
-	HWAccel          string      `yaml:"hw_accel"`
-	VideoBitRate     string      `yaml:"video_bit_rate"`
-	VideoQuality     int         `yaml:"video_quality"`
-	Preset           string      `yaml:"preset"`
-	Scale            string      `yaml:"scale"`
-	KeyframeInterval int         `yaml:"keyframe_interval"`
-	Muxer            string      `yaml:"muxer"`
+// Task _
+type Task struct {
+	ID      string `yaml:"id"`
+	InFile  string `yaml:"in_file"`
+	OutFile string `yaml:"out_file"`
+	Params  Params
+}
+
+// Params _
+type Params struct {
+	VideoCodec       string `yaml:"video_codec"`
+	HWAccel          string `yaml:"hw_accel"`
+	VideoBitRate     string `yaml:"video_bit_rate"`
+	VideoQuality     int    `yaml:"video_quality"`
+	Preset           string `yaml:"preset"`
+	Scale            string `yaml:"scale"`
+	KeyframeInterval int    `yaml:"keyframe_interval"`
 }
 
 // ErrFileIsNotVideo _
@@ -72,38 +75,3 @@ var ErrOutputFileExistsOrIsDirectory = errors.New("Output file exists or is dire
 
 // ErrVtbQualityNotSupported _
 var ErrVtbQualityNotSupported = errors.New("Video quality option is not supported by Apple VideoToolBox")
-
-// UnmarshalYAML _
-func (ct *ConverterTask) UnmarshalYAML(unmarshal func(interface{}) error) error {
-
-	task := struct {
-		ID               string `yaml:"id"`
-		InFile           string `yaml:"in_file"`
-		OutFile          string `yaml:"out_file"`
-		VideoCodec       string `yaml:"video_codec"`
-		HWAccel          string `yaml:"hw_accel"`
-		VideoBitRate     string `yaml:"video_bit_rate"`
-		VideoQuality     int    `yaml:"video_quality"`
-		Preset           string `yaml:"preset"`
-		Scale            string `yaml:"scale"`
-		KeyframeInterval int    `yaml:"keyframe_interval"`
-	}{}
-
-	if err := unmarshal(&task); err != nil {
-		return err
-	}
-
-	ct.ID = task.ID
-	ct.VideoCodec = task.VideoCodec
-	ct.HWAccel = task.HWAccel
-	ct.VideoBitRate = task.VideoBitRate
-	ct.VideoQuality = task.VideoQuality
-	ct.Preset = task.Preset
-	ct.Scale = task.Scale
-	ct.KeyframeInterval = task.KeyframeInterval
-
-	ct.InFile = files.NewFile(task.InFile)
-	ct.OutFile = files.NewFile(task.OutFile)
-
-	return nil
-}