+++ pkg/media/convert/nvenc_hwaccel.go
@@ -6,7 +6,7 @@ import (
 )
 
 type nvencHWAccel struct {
-	task     ConverterTask
+	task     Task
 	metadata ffmpegModels.Metadata
 }
 
@@ -32,13 +32,13 @@ func (hw *nvencHWAccel) configure(mediaFile *ffmpegModels.Mediafile) error {
 
 	mediaFile.SetNvencRateControl("constqp")
 
-	if hw.task.VideoQuality > 0 {
+	if hw.task.Params.VideoQuality > 0 {
 		mediaFile.SetCRF(0)
 		mediaFile.SetLibx265Params(nil)
-		mediaFile.SetConstantQuantization(hw.task.VideoQuality)
+		mediaFile.SetConstantQuantization(hw.task.Params.VideoQuality)
 	}
 
-	switch hw.task.VideoCodec {
+	switch hw.task.Params.VideoCodec {
 	case HevcCodecType:
 		mediaFile.SetVideoCodec("hevc_nvenc")
 	case H264CodecType: